Move async trainer to run_experiment.
