--- dopamine/discrete_domains/run_experiment/helper.py
@@ -1,107 +1,0 @@
-# coding=utf-8
-# Copyright 2018 Google Inc. All Rights Reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#      http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-"""Module defining and methods to create and train an agent."""
-
-from __future__ import absolute_import
-from __future__ import division
-from __future__ import print_function
-
-from dopamine.agents.dqn import dqn_agent
-from dopamine.agents.implicit_quantile import implicit_quantile_agent
-from dopamine.agents.rainbow import rainbow_agent
-from dopamine.discrete_domains.run_experiment import run_async_training
-from dopamine.discrete_domains.run_experiment import run_experiment
-
-import gin.tf
-
-
-def load_gin_configs(gin_files, gin_bindings):
-  """Loads gin configuration files.
-
-  Args:
-    gin_files: list, of paths to the gin configuration files for this
-      experiment.
-    gin_bindings: list, of gin parameter bindings to override the values in
-      the config files.
-  """
-  gin.parse_config_files_and_bindings(gin_files,
-                                      bindings=gin_bindings,
-                                      skip_unknown=False)
-
-
-@gin.configurable
-def create_agent(sess, environment, agent_name=None, summary_writer=None,
-                 debug_mode=False):
-  """Creates an agent.
-
-  Args:
-    sess: A `tf.Session` object for running associated ops.
-    environment: An Atari 2600 Gym environment.
-    agent_name: str, name of the agent to create.
-    summary_writer: A Tensorflow summary writer to pass to the agent
-      for in-agent training statistics in Tensorboard.
-    debug_mode: bool, whether to output Tensorboard summaries. If set to true,
-      the agent will output in-episode statistics to Tensorboard. Disabled by
-      default as this results in slower training.
-
-  Returns:
-    agent: An RL agent.
-
-  Raises:
-    ValueError: If `agent_name` is not in supported list.
-  """
-  assert agent_name is not None
-  if not debug_mode:
-    summary_writer = None
-  if agent_name == 'dqn':
-    return dqn_agent.DQNAgent(sess, num_actions=environment.action_space.n,
-                              summary_writer=summary_writer)
-  elif agent_name == 'rainbow':
-    return rainbow_agent.RainbowAgent(
-        sess, num_actions=environment.action_space.n,
-        summary_writer=summary_writer)
-  elif agent_name == 'implicit_quantile':
-    return implicit_quantile_agent.ImplicitQuantileAgent(
-        sess, num_actions=environment.action_space.n,
-        summary_writer=summary_writer)
-  else:
-    raise ValueError('Unknown agent: {}'.format(agent_name))
-
-
-@gin.configurable
-def create_runner(base_dir, schedule='continuous_train_and_eval'):
-  """Creates an experiment Runner.
-
-  Args:
-    base_dir: str, base directory for hosting all subdirectories.
-    schedule: string, which type of Runner to use.
-
-  Returns:
-    runner: A `Runner` like object.
-
-  Raises:
-    ValueError: When an unknown schedule is encountered.
-  """
-  assert base_dir is not None
-  # Continuously runs training and evaluation until max num_iterations is hit.
-  if schedule == 'continuous_train_and_eval':
-    return run_experiment.Runner(base_dir, create_agent)
-  # Continuously runs training until max num_iterations is hit.
-  elif schedule == 'continuous_train':
-    return run_experiment.TrainRunner(base_dir, create_agent)
-  elif schedule == 'async_train':
-    return run_async_training.AsyncRunner(base_dir, create_agent)
-  else:
-    raise ValueError('Unknown schedule: {}'.format(schedule))
